affilio_mcp/requirements.txt
fastapi[standard]
uvicorn
odmantic
motor
redis
pillow
fastmcp
mcp[cli]
python-dateutil
segno
tldextract
backoff
gunicorn
